Remove setting for "Twist And Go". Place it in the debug panel. [#2921]
