+++ assets/strings.en.json
@@ -1662,15 +1662,11 @@
     "panel.settings.icard.bluetooth.sensitivity.normal.label": "Normal",
     "panel.settings.icard.bluetooth.sensitivity.low.label": "Low",
     "panel.settings.icard.vibrate.button": "Vibrate when unlocking",
-    "panel.settings.icard.twist_n_go.title.label": "Twist And Go",
-    "panel.settings.icard.twist_n_go.description.label": "Rotate your mobile device 90\u00B0 to the right and left, as if turning a door knob.",
-    "panel.settings.icard.twist_n_go.description2.label": "Doors must be enabled for Twist and Go to work.",
     "panel.settings.icard.ios_settings.title.label": "Open System Settings",
     "panel.settings.icard.ios_settings.description.label": "Return to the {{app_title}} app by tapping the return arrow at top of the page.",
     "panel.settings.icard.notification_drawer.title.label": "Notification ???",
     "panel.settings.icard.notification_drawer.description.label": "Unlock doors from notification drawer.",
     "panel.settings.icard.rssi.change.failed.msg": "Failed to change Bluetooth sensitivity.",
-    "panel.settings.icard.twist_n_go.change.failed.msg": "Failed to %s Twist And Go. Please, try again later.",
     "panel.settings.icard.unlock.vibration.change.failed.msg": "Failed to %s vibration when unlocking. Please, try again later.",
     "panel.settings.icard.unlock.sound.change.failed.msg": "Failed to %s sound when unlocking. Please, try again later.",
     "panel.settings.icard.enable.label": "enable",
