+++ assets/strings.es.json
@@ -1660,15 +1660,11 @@
     "panel.settings.icard.bluetooth.sensitivity.normal.label": "Normal",
     "panel.settings.icard.bluetooth.sensitivity.low.label": "Bajo",
     "panel.settings.icard.vibrate.button": "Vibrar al desbloquear",
-    "panel.settings.icard.twist_n_go.title.label": "Gira y vete",
-    "panel.settings.icard.twist_n_go.description.label": "Gira tu dispositivo móvil 90\u00B0 hacia la derecha y hacia la izquierda, como si giraras el pomo de una puerta.",
-    "panel.settings.icard.twist_n_go.description2.label": "Las puertas deben estar habilitadas para que Twist and Go funcione.",
     "panel.settings.icard.ios_settings.title.label": "Abrir configuración del sistema",
     "panel.settings.icard.ios_settings.description.label": "Vuelva a la aplicación {{app_title}} tocando la flecha de retorno en la parte superior de la página.",
     "panel.settings.icard.notification_drawer.title.label": "Notificación ???",
     "panel.settings.icard.notification_drawer.description.label": "Desbloquee las puertas del cajón de notificaciones.",
     "panel.settings.icard.rssi.change.failed.msg": "No se pudo cambiar la sensibilidad de Bluetooth.",
-    "panel.settings.icard.twist_n_go.change.failed.msg": "Error al %s Twist And Go. Por favor, inténtelo de nuevo más tarde.",
     "panel.settings.icard.unlock.vibration.change.failed.msg": "No se pudo %s vibrar al desbloquear. Por favor, inténtelo de nuevo más tarde.",
     "panel.settings.icard.unlock.sound.change.failed.msg": "No se pudo escuchar el sonido %s al desbloquear. Por favor, inténtelo de nuevo más tarde.",
     "panel.settings.icard.enable.label": "permitir",
