+++ assets/strings.zh.json
@@ -1662,15 +1662,11 @@
 		"panel.settings.icard.bluetooth.sensitivity.normal.label": "普通的",
 		"panel.settings.icard.bluetooth.sensitivity.low.label": "低的",
 		"panel.settings.icard.vibrate.button": "解鎖時振動",
-		"panel.settings.icard.twist_n_go.title.label": "扭轉和去",
-		"panel.settings.icard.twist_n_go.description.label": "將您的移動設備向右和向左旋轉 90\u00B0，就像轉動門把手一樣。",
-		"panel.settings.icard.twist_n_go.description2.label": "必須啟用門才能使 Twist and Go 工作。",
 		"panel.settings.icard.ios_settings.title.label": "打開系統設置",
 		"panel.settings.icard.ios_settings.description.label": "點擊頁面頂部的返回箭頭返回 {{app_title}} 應用程序。",
 		"panel.settings.icard.notification_drawer.title.label": "通知 ？？？",
 		"panel.settings.icard.notification_drawer.description.label": "從通知抽屜解鎖門。",
 		"panel.settings.icard.rssi.change.failed.msg": "無法更改藍牙靈敏度。",
-		"panel.settings.icard.twist_n_go.change.failed.msg": "%s Twist And Go 失敗。 請稍後再試。",
 		"panel.settings.icard.unlock.vibration.change.failed.msg": "解鎖時 %s 振動失敗。 請稍後再試。",
 		"panel.settings.icard.unlock.sound.change.failed.msg": "解鎖時無法發出 %s 聲音。 請稍後再試。",
 		"panel.settings.icard.enable.label": "使能夠",
